Refactor epicenter to project status tab
