+++ client/src/App.tsx
@@ -21,6 +21,7 @@ import CategoryGroup from "./components/categoryGroup/CategoryGroup";
 import AdminRoute from "./util/AdminRoute";
 import JudgeRoute from "./util/JudgeRoute";
 import SponsorRoute from "./util/SponsorRoute";
+import ProjectStatusHome from "./components/projectStatus/ProjectStatusHome";
 
 const { Content } = Layout;
 
@@ -44,12 +45,23 @@ function App() {
             <Switch>
               <Route exact path="/" render={() => <Dashboard user={data} />} />
               <Route exact path="/create" render={() => <SubmissionFormContainer user={data} />} />
-              <SponsorRoute exact path="/category-group/:categoryGroupId" component={CategoryGroup} user={data}/>
+              <SponsorRoute
+                exact
+                path="/category-group/:categoryGroupId"
+                component={CategoryGroup}
+                user={data}
+              />
               <Route exact path="/projectgallery" render={() => <ProjectGallery user={data} />} />
               <Route exact path="/projects/:projectId" component={ProjectDetails} />
-              <JudgeRoute exact path="/judging" render={() => <JudgingHome user={data} />} user={data}/>
-              <AdminRoute exact path="/admin/:activePane?" component={AdminHome} user={data}/>
-              <AdminRoute exact path="/epicenter" component={Epicenter} user={data}/>
+              <JudgeRoute
+                exact
+                path="/judging"
+                render={() => <JudgingHome user={data} />}
+                user={data}
+              />
+              <AdminRoute exact path="/admin/:activePane?" component={AdminHome} user={data} />
+              <AdminRoute exact path="/epicenter" component={Epicenter} user={data} />
+              <AdminRoute exact path="/project-status" component={ProjectStatusHome} user={data} />
               <Route component={NotFoundDisplay} />
             </Switch>
           </Content>
